pyproject.toml
[tool.poetry]
name = "ugc-sprint-1"
version = "1.0.0"
description = ""
authors = [
    "Носов Сергей <SONosov@datafort.ru>",
    "Бубнова Яна <bujhvh@mail.ru>",
    "Ротанин Дмитрий <SONosov@datafort.ru>",
]
readme = "README.md"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dependencies]
python = "~3.11.9"
fastapi = "^0.112.1"
aiokafka = "^0.11.0"
gunicorn = "^23.0.0"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
clickhouse-driver = "^0.2.9"
uvicorn = "^0.30.6"
requests = "^2.32.3"
pyjwt = "2.8.0"
backoff = "^2.2.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"
pytest-env = "^1.1.3"
pytest-mock = "^3.14.0"
pytest-randomly = "^3.15.0"
pytest-icdiff = "^0.9"
httpx = "^0.27.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.8.0"
flake8 = "^7.1.1"
flake8-pyproject = "^1.2.3"
bandit = "^1.7.9"
mypy = "^1.11.1"


[tool.isort]
profile = 'black'


[tool.black]
line-length = 120
skip-string-normalization = true


[tool.flake8]
exclude = [".git", "__pycache__", ".venv"]
max-line-length = 120


[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
plugins = [
    "pydantic.mypy"
]
